Fix after refactoring
