+++ app/frontend/components/tab_search/tab_search.js
@@ -19,7 +19,7 @@
 const VerseSearch = require('./verse_search.js');
 const { waitUntilIdle } = require('../../helpers/ezra_helper.js');
 const eventController = require('../../controllers/event_controller.js');
-const verseListController = require('../controllers/verse_list_controller.js');
+const verseListController = require('../../controllers/verse_list_controller.js');
 
 /**
  * The TabSearch component implements the in-tab search functionality which is enabled with CTRL + f / CMD + f.
